// src/User/interfaces/user.interface.ts
/* eslint-disable prettier/prettier */
export interface UserInterface {
  uuid: string;
  firstname: string;
  lastname: string;
  password: string;
  contact: string;
  email: string;
  entreprise_id: string;
  voiture: string;
  statut: string;
  photo: string;
  active: string;
  created_at: string;
  updated_at: string;
  role: string;
}